refactor(frontend): rename all ETH transactions components
