+++ src/frontend/src/eth/components/transactions/EthTransactionsSkeletons.svelte
